requirements_dev.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=requirements_dev.txt requirements_dev.in
#
amqp==5.1.1
    # via
    #   -r ./requirements_test.txt
    #   kombu
asgiref==3.6.0
    # via
    #   -r ./requirements_test.txt
    #   django
attrs==22.2.0
    # via
    #   -r ./requirements_test.txt
    #   pytest
billiard==3.6.4.0
    # via
    #   -r ./requirements_test.txt
    #   celery
build==0.10.0
    # via pip-tools
cachetools==5.3.0
    # via
    #   -r ./requirements_test.txt
    #   tox
celery==5.2.7
    # via
    #   -r ./requirements_test.txt
    #   django-celery-beat
    #   django-celery-results
certifi==2022.12.7
    # via
    #   -r ./requirements_test.txt
    #   requests
chardet==5.1.0
    # via
    #   -r ./requirements_test.txt
    #   tox
charset-normalizer==3.0.1
    # via
    #   -r ./requirements_test.txt
    #   requests
click==8.1.3
    # via
    #   -r ./requirements_test.txt
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
    #   uvicorn
click-didyoumean==0.3.0
    # via
    #   -r ./requirements_test.txt
    #   celery
click-plugins==1.1.1
    # via
    #   -r ./requirements_test.txt
    #   celery
click-repl==0.2.0
    # via
    #   -r ./requirements_test.txt
    #   celery
colorama==0.4.6
    # via
    #   -r ./requirements_test.txt
    #   tox
coverage[toml]==7.0.5
    # via
    #   -r ./requirements_test.txt
    #   pytest-cov
distlib==0.3.6
    # via
    #   -r ./requirements_test.txt
    #   virtualenv
django==4.1.5
    # via
    #   -r ./requirements_test.txt
    #   django-celery-beat
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   djangorestframework
django-celery-beat==2.4.0
    # via -r ./requirements_test.txt
django-celery-results==2.4.0
    # via -r ./requirements_test.txt
django-debug-toolbar==3.8.1
    # via -r requirements_dev.in
django-extensions==3.2.1
    # via -r ./requirements_test.txt
django-filter==22.1
    # via -r ./requirements_test.txt
django-simple-history==3.2.0
    # via -r ./requirements_test.txt
django-timezone-field==5.0
    # via
    #   -r ./requirements_test.txt
    #   django-celery-beat
djangorestframework==3.14.0
    # via -r ./requirements_test.txt
exceptiongroup==1.1.0
    # via
    #   -r ./requirements_test.txt
    #   pytest
factory-boy==3.2.1
    # via -r ./requirements_test.txt
faker==16.6.0
    # via
    #   -r ./requirements_test.txt
    #   factory-boy
filelock==3.9.0
    # via
    #   -r ./requirements_test.txt
    #   tox
    #   virtualenv
flake8==6.0.0
    # via -r ./requirements_test.txt
freezegun==1.2.2
    # via -r ./requirements_test.txt
gunicorn==20.1.0
    # via -r ./requirements_test.txt
h11==0.14.0
    # via
    #   -r ./requirements_test.txt
    #   uvicorn
idna==3.4
    # via
    #   -r ./requirements_test.txt
    #   requests
    #   yarl
iniconfig==2.0.0
    # via
    #   -r ./requirements_test.txt
    #   pytest
isort==5.11.4
    # via -r ./requirements_test.txt
kombu==5.2.4
    # via
    #   -r ./requirements_test.txt
    #   celery
mccabe==0.7.0
    # via
    #   -r ./requirements_test.txt
    #   flake8
multidict==6.0.4
    # via
    #   -r ./requirements_test.txt
    #   yarl
packaging==23.0
    # via
    #   -r ./requirements_test.txt
    #   build
    #   pyproject-api
    #   pytest
    #   tox
pip-tools==6.12.1
    # via -r requirements_dev.in
platformdirs==2.6.2
    # via
    #   -r ./requirements_test.txt
    #   tox
    #   virtualenv
pluggy==1.0.0
    # via
    #   -r ./requirements_test.txt
    #   pytest
    #   tox
prompt-toolkit==3.0.36
    # via
    #   -r ./requirements_test.txt
    #   click-repl
psycopg2==2.9.5
    # via -r ./requirements_test.txt
pycodestyle==2.10.0
    # via
    #   -r ./requirements_test.txt
    #   flake8
pyflakes==3.0.1
    # via
    #   -r ./requirements_test.txt
    #   flake8
pyproject-api==1.5.0
    # via
    #   -r ./requirements_test.txt
    #   tox
pyproject-hooks==1.0.0
    # via build
pytest==7.2.1
    # via
    #   -r ./requirements_test.txt
    #   pytest-cov
    #   pytest-django
pytest-cov==4.0.0
    # via -r ./requirements_test.txt
pytest-django==4.5.2
    # via -r ./requirements_test.txt
python-crontab==2.7.1
    # via
    #   -r ./requirements_test.txt
    #   django-celery-beat
python-dateutil==2.8.2
    # via
    #   -r ./requirements_test.txt
    #   faker
    #   freezegun
    #   python-crontab
pytz==2022.7.1
    # via
    #   -r ./requirements_test.txt
    #   celery
    #   django-timezone-field
    #   djangorestframework
pyyaml==6.0
    # via
    #   -r ./requirements_test.txt
    #   vcrpy
requests==2.28.2
    # via -r ./requirements_test.txt
six==1.16.0
    # via
    #   -r ./requirements_test.txt
    #   click-repl
    #   python-dateutil
    #   vcrpy
sqlparse==0.4.3
    # via
    #   -r ./requirements_test.txt
    #   django
    #   django-debug-toolbar
tomli==2.0.1
    # via
    #   -r ./requirements_test.txt
    #   build
    #   coverage
    #   pyproject-api
    #   pytest
    #   tox
tox==4.3.5
    # via -r ./requirements_test.txt
tzdata==2022.7
    # via
    #   -r ./requirements_test.txt
    #   django-celery-beat
urllib3==1.26.14
    # via
    #   -r ./requirements_test.txt
    #   requests
uvicorn==0.20.0
    # via -r ./requirements_test.txt
vcrpy==4.2.1
    # via -r ./requirements_test.txt
vine==5.0.0
    # via
    #   -r ./requirements_test.txt
    #   amqp
    #   celery
    #   kombu
virtualenv==20.17.1
    # via
    #   -r ./requirements_test.txt
    #   tox
wcwidth==0.2.6
    # via
    #   -r ./requirements_test.txt
    #   prompt-toolkit
wheel==0.38.4
    # via pip-tools
whitenoise==6.3.0
    # via -r ./requirements_test.txt
wrapt==1.14.1
    # via
    #   -r ./requirements_test.txt
    #   vcrpy
yarl==1.8.2
    # via
    #   -r ./requirements_test.txt
    #   vcrpy

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
